fix: [진건오] 일반회원 장바구니 상품 삭제 시스템[DIY-SALES-PROCESS 24]

+++ spring/buy_idea/src/main/java/team_project/buy_idea/service/order/shopppingbucket/ShoppingBucketServiceImpl.java
@@ -12,7 +12,6 @@ import team_project.buy_idea.repository.member.MemberRepository;
 import team_project.buy_idea.repository.order.shopppingBucket.ShoppingBucketProductRepository;
 import team_project.buy_idea.repository.order.shopppingBucket.ShoppingBucketRepository;
 import team_project.buy_idea.repository.product.ProductRepository;
-import team_project.buy_idea.service.security.RedisService;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,8 +20,6 @@ import java.util.Optional;
 @Service
 public class ShoppingBucketServiceImpl implements ShoppingBucketService{
 
-    @Autowired
-    private RedisService redisService;
     @Autowired
     private ShoppingBucketRepository shoppingBucketRepository;
 
@@ -37,8 +34,6 @@ public class ShoppingBucketServiceImpl implements ShoppingBucketService{
 
     @Override
     public void addProductToShoppingBucket(ShoppingBucketRequest bucketRequest) {
-//        String memberToken = bucketRequest.getMemberToken();
-//        Long memberId = redisService.getValueByKey(memberToken);
 
         String nickname = bucketRequest.getNickname();
 
@@ -83,7 +78,38 @@ public class ShoppingBucketServiceImpl implements ShoppingBucketService{
     }
 
     @Override
-    public void deleteShoppingBucketProduct(Long itemId) {
-        bucketProductRepository.deleteById(itemId);
+    public void deleteShoppingBucketProduct(Long itemId, String nickname) {
+        List<ShoppingBucketItem> shoppingBucketItems = this.shoppingBucketItemList(nickname);
+
+        if (shoppingBucketItems.isEmpty()){
+            Optional<Member> maybeMember = memberRepository.findBuyDiaMemberByNickname(nickname);
+            Member member = maybeMember.get();
+
+            ShoppingBucket shoppingBucket = ShoppingBucket.builder()
+                    .member(member)
+                    .shoppingBucketTotalCnt(0)
+                    .build();
+
+            shoppingBucketRepository.save(shoppingBucket);
+        } else {
+            bucketProductRepository.deleteById(itemId);
+            List<ShoppingBucketItem> shoppingBucketItemList = this.shoppingBucketItemList(nickname);
+
+            if (shoppingBucketItemList.size() >= 0){
+                Optional<ShoppingBucket> maybeShoppingBucket = shoppingBucketRepository.findByNickname(nickname);
+                ShoppingBucket shoppingBucket = maybeShoppingBucket.get();
+
+                shoppingBucket.setShoppingBucketTotalCnt(shoppingBucket.getShoppingBucketTotalCnt() -1);
+
+                shoppingBucketRepository.save(shoppingBucket);
+            } else {
+                Optional<ShoppingBucket> maybeShoppingBucket = shoppingBucketRepository.findByNickname(nickname);
+                ShoppingBucket shoppingBucket = maybeShoppingBucket.get();
+
+                shoppingBucket.setShoppingBucketTotalCnt(0);
+
+                shoppingBucketRepository.save(shoppingBucket);
+            }
+        }
     }
 }
